requirements.txt
flask==2.0.2
Werkzeug==2.0.2

# Windows

pywebview
cryptography

# Android

kivy[base]
cython
jnius
